Moved vector functions to utilities
